Add Store variant separation
