+++ mobile/src/HuaweiAppGallery/java/com/murati/audiobook/utils/AdHelper.java
@@ -17,6 +17,8 @@ import com.murati.audiobook.R;
 
 import java.util.Locale;
 
+import static com.murati.audiobook.utils.MobileServicesHelper.isHmsAvailable;
+
 public class AdHelper {
 
     private static final String TAG = LogHelper.makeLogTag(AdHelper.class);
@@ -64,11 +66,6 @@ public class AdHelper {
 
     public static void tryLoadAds(AppCompatActivity activity, String tag) {
         try {
-            AdHelper.loadGoogleAdmodToView(activity, R.id.googleAdView);
-        } catch (Exception ex) {
-            Log.e(tag, ex.getMessage());
-        }
-        try {
             AdHelper.loadHuaweiAdkitToView(activity, R.id.huaweiAdView);
         } catch (Exception ex) {
             Log.e(tag, ex.getMessage());
@@ -89,20 +86,6 @@ public class AdHelper {
         AdParam adParam = new AdParam.Builder().build();
         huaweiAdView.loadAd(adParam);
         huaweiAdView.setVisibility(View.VISIBLE);
-    }
-    private static void loadGoogleAdmodToView(AppCompatActivity activity, int resourceId) {
-        MobileAds.initialize(activity, BuildConfig.ADMOB_APP_ID);
-        AdView adView = activity.findViewById(resourceId);
-
-        // Admob testing: https://developers.google.com/admob/android/test-ads
-        // if (BuildConfig.DEBUG) // TEST banner
-        //     adView.setAdUnitId("ca-app-pub-3940256099942544/6300978111");
-        // else
-        //     adView.setAdUnitId(BuildConfig.ADMOB_UNIT_ID_1);
-
-        AdRequest adRequest = new AdRequest.Builder().build();
-        adView.loadAd(adRequest);
-        adView.setVisibility(View.VISIBLE);
     }
 
     public static int getAdPosition(FirebaseRemoteConfig mFirebaseRemoteConfig) {
